pyproject.toml
[project]
name = "serial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "nicegui>=2.12.1",
    "pyserial>=3.5",
]
